+++ R/ropenblas.R
@@ -774,7 +774,7 @@ compiler_r <- function(r_version = NULL,
     answer <-
       "{cli::col_blue(cli::symbol$bullet)} R version already compiled: (yes - changes without recompiling) and (no - compiles again)"  %>%
       glue %>%
-      answer_yes_no
+      answer_yes_no()
     
     validate_answer(answer)
     
@@ -791,11 +791,11 @@ compiler_r <- function(r_version = NULL,
   download <- download_r(x = r_version)
   
   if (is.null(with_blas)) {
-    with_blas <-  "-L/opt/OpenBLAS/lib \\
+    with_blas <-  "-L/opt/OpenBLAS/lib -lopenblas \\
      -I/opt/OpenBLAS/include \\
      -lpthread \\
      -lm" %>% 
-      glue
+      glue(.)
   }
   
   if (is.null(complementary_flags))
@@ -808,8 +808,8 @@ compiler_r <- function(r_version = NULL,
      --enable-R-shlib \\
      --enable-threads=posix \\
      --with-blas=\"{with_blas}\" \\
-     {complementary_flags}" %>%  
-    glue
+     {complementary_flags}" %>%
+    glue()
   
   if (dir.exists("/opt/OpenBLAS/lib/") && dir_blas()$use_openblas) {
     # configure ---------------------------------------------------------------
@@ -870,6 +870,16 @@ compiler_r <- function(r_version = NULL,
       new = download,
       code = run_command(x = "make install PREFIX=/opt/R/{r_version}", key_root = key_root)
     )
+    
+    # creating symbolic links -------------------------------------------------
+    
+    "ln -sf /opt/R/{r_version}/bin/R {dir_r}"  %>%
+      glue %>%
+      run_command(key_root = key_root)
+    
+    "ln -sf /opt/R/{r_version}/bin/Rscript {dir_rscript}" %>%
+      glue %>%
+      run_command(key_root = key_root)
     
     fix_openblas_link(restart_r = FALSE, key_root = key_root)
     
@@ -1016,7 +1026,7 @@ link_again <- function(restart_r = TRUE) {
   if (dir_blas()$use_openblas) {
     "{cli::col_blue(cli::symbol$bullet)} Linking again is not necessary. {style_underline(style_bold(\"R\"))} \\
       already uses the {style_underline(style_bold(\"OpenBLAS\"))} library. You can stay calm." %>%
-      glue
+      glue()
   } else {
     if (!dir.exists("/opt/OpenBLAS/lib"))
       "{cli::col_blue(cli::symbol$bullet)} Run the {style_underline(style_bold(\"ropenblas()\"))} function ..." %>%